+++ dmsapi/api/point.py
@@ -1,5 +1,0 @@
-from dmsapi.api.base import BaseAPI
-
-
-class Point(BaseAPI):
-    pass